Added a way to typeset arbitrary LaTeX using jsmath, also provided a workaround to a maxima LaTeX bug.
Added functionality to typeset general SAGE objects using jsmath, as well as
using jsmath to render arbitray strings of LaTeX markup.

Also parsed the string that maxima returns when you request it to TeX a Maxima
object. Maxima had a nasty bug that represented (in TeX) the multiplication of
two integer literals by a space between them. The workaround sub's such
whitespace with a \cdot.