cython option to make inplace operations safe
requires Cython >= 0.9.6.6

As of the current count, this will add 1644 new increfs to the library (compared to 27843 total).
That doesn't say too much, as what matters are which ones are in inner loops, etc.
but enabling inplace operations should more than make up for this.

The issue this fixes is that local variables in cdef'd functions have one less refcount
than one would expect.