#5449 Implement a variant of @cached_method with cache stored in the parent
When the elements of a parent do not have unique representation, it
can be desirable to store the cache of (some of) the methods in the
parent rather than in the element.

This patches implements a variant @cached_in_parent_method of the
decorator @cached_method which does just this.