For #5929: fix timeout in maxima/ecl due to out-of-synch condition
This is triggered by the following doctest
{{{
sage: maxima._sendstr('1/1'*500)
sage: maxima('2+2')
}}}
In fact, the first line is missing a semicolon, so maxima is stuck out
of synch; the synchronization code is run for the second line, and
after a timeout, it tries to break by sending a ctrl-c (this happens
in Expect._interrupt()), but ctrl-c is broken with ecl, so this hangs.

The current fix adds an explicit semicolon to the synchronization
code. This fixes the timeout in the above doctest. The hang is still
latent, because there seems to be no way to interrupt maxima/ecl.