#6088: LatticePolytope: Removed a try/catch which could involuntarily hide exceptions from lower code
Removed a try/catch which could involuntarily hide exceptions from lower code.

This try catch made it hard to discover a trivial bug in Sequences
because it was catching the corresponding exception.

Its purpose was just to test if some data had already been
cached. This is not on a critical section, so testing on the existence
of an attribute seems as good, while clearer and safer.

By the way, I would recommend not to use Sequence for this kind of
applications, as the overhead in speed and complexity is non trivial,
whereas the specific features of Sequence do not seem to be very much
used here. Plain tuples should probably work as well.