#6097: Implements a mantra for declaring abstract methods
This patch implements a decorator tha can be used to declare a method
that should be implemented by derived classes. This declaration should
typically include documentation for the specification for this method.

The purpose is to enforce a consistent and visual mantra for such
declarations. This is also used by #5891 for automated tests.