#8444: fix memory leak due to dup call to _sig_on in the bottom part of PariInstance.__call__
At the bottom of PariInstance.__call__ both _sig_on (first) and _sig_str
(later) are used. In fact, both count as calls to _sig_on (actually _sig_on is
equivalent to _sig_str(NULL)) and these are ''not'' reentrant, i.e. nesting is not
supported (anyway, there's only one implicit _sig_off in the call to new_gen).

A double _sig_on, as defined in interrupt.h, is usually equivalent to a single
one -- however, for the pari library these macros are overrided as _pari_sig_on
(defined in misc.h and pari_err.h) adding a call to pari's own error catching
function.  Calling the err_catch() function twice without the corresponding
call to err_leave() results in a memory leak which is reported in #8444.

The patch is one-liner: removing the first _sig_on fixes the memory leak.

Note: this line was added by changeset 10536:423520e7d069 as part of a large
effort to add lots of missing calls to _sig_off in pari interface. However, I
think in this case it was just an oversight because the already existing call to
_sig_on was only implicit in the call to _sig_str.