Trac 8963: Make tableau row_stabilizer use all elements of the tableau
The row_stabilizer method of a tableau ignores elements of the tableau
which are alone in their row.  This patch ensures that each element of
the resulting row_stabilizer group is a permutation of the set {1, ..., k},
where k is the largest entry of the tableau.

It also creates a new method entries() which returns the list of
entries of a tableau, something that I need quite regularly.