Trac #10446: trac #10446: Implementation of the Schuetzenberger involution, the Lusztig involution and the promotion operator.
This patch implements the Schuetzenberger involution on both words and tableaux.
In addition it generalizes the promotion operator on tableaux to act not only
on rectangular tableaux, but tableaux of any shape.

For crystals in the category ClassicalCrystals, the Lusztig involution is implemented and
tested in type `A` to give the same result as the Schuetzenberger involution.
A section about the Lusztig involution is added to the thematic tutorial on crystals.