Trac #11597: Trac #11597: Dimension of the singular set
This patch adds a method dimension_singularities()
to toric varieties that computes the dimension of the singular
set. This is usually one of the first questions one has, so I
find it rather convenient to have a method for it. It also adds
an optional fan.is_smooth(codim=d) argument for expert use,
though in practice I think its easy to get confused with
dimension and codimension so I did not add a similar optional
argument to toric_variety.is_smooth().