Trac #12471: Trac 12471: Fix numerical evaluation of symbolic expressions with complex coefficients.
When converting a symbolic expression to a real domain like float, RR or RDF,
the _eval_self() function first tries to evaluate each coefficient in the
expression in this domain. This breaks if the coefficients are complex, but
the result is still real:

sage: cos(I).n()
1.54308063481524
sage: RR(cos(I))
<boom>

In this case, we try the evaluation once more with the complex_field()
corresponding to the given domain.