Trac #8899: and #11929: Implement noncommutative symmetric functions and quasi-symmetric functions
Noncommutative symmetric functions and quasi-symmetric functions. Each algebra
is implemented as a Hopf algebra with realizations (the realizations being the
various bases of the algebras).

* Bases implemented for NCSF, and change of bases between them:

    - Complete
    - Ribbon
    - Elementary
    - Psi (power sums)
    - Phi (power sums)

* Bases implemented for QSym, and change of bases between them:

    - Monomial
    - Fundamental

There is also a method a_realization that returns a particular realization
of the algebra. Computations that are not yet implemented in basis are
performed by converting to a_realization(). Current implementation:

    - NCSF.a_realization() returns the Complete basis
    - QSym.a_realization() returns the Monomial basis

HopfAlgebras.Realizations.ParentMethods.antipode_by_coercion offers a default
implementation that converts a_realization in order to do the computation. This
is used by the antipode method as a last resort if no other implementation is
found.

Added extra doctests for NCSF and QSym for default the counit and antipode
methods in Coalgebras.(With)Realizations.